Separating FCGR/DFT directories and cleaning up
