+++ DFT/scripts/download_genomes.sh
