+++ DFT/scripts/euclidean_dist.py
