+++ DFT/scripts/windowed_CGR.py
