--- scripts/.Rhistory
@@ -1,158 +1,0 @@
-library(readr)
-FCGRs_for_R <- read_delim("~/PycharmProjects/Master/test_files/FCGRs_for_R",
-"\t", escape_double = FALSE, trim_ws = TRUE)
-View(FCGRs_for_R)
-FCGRs_for_R <- read_delim("~/PycharmProjects/Master/test_files/FCGRs_for_R", "\t")
-FCGRs_for_R
-View(FCGRs_for_R)
-test <- t(FCGRs_for_R)
-View(FCGRs_for_R)
-View(test)
-cor_test <- cor(test)
-View(cor_test)
-R <- cor(test)
-eigen(R)$values
-eigen(R)$vectors
-R <- var(test)
-#Eigenvalues :
-val.p <- eigen(R)$values
-#Eigenvectors :
-vec.p <- eigen(R)$vectors
-sum(val.p[1:2])/sum(val.p) # % de variance totale expliquée par 2 premiers vecteurs propres
-U <- vec.p
-U
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space")
-dat <- read_delim("~/PycharmProjects/Master/Main/data/genomes/e_coli/FCGRs_for_R", "\t")
-View(dat)
-test <- t(FCGRs_for_R)
-R <- var(test)
-#Eigenvalues :
-val.p <- eigen(R)$values
-#Eigenvectors :
-vec.p <- eigen(R)$vectors
-sum(val.p[1:2])/sum(val.p) # % de variance totale expliquée par 2 premiers vecteurs propres
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space")
-dat <- read_delim("~/PycharmProjects/Master/Main/data/genomes/e_coli/FCGRs_for_R", "\t")
-tdat <- t(dat)
-R <- var(tdat)
-#Eigenvalues :
-val.p <- eigen(R)$values
-#Eigenvectors :
-vec.p <- eigen(R)$vectors
-sum(val.p[1:2])/sum(val.p) # % de variance totale expliquée par 2 premiers vecteurs propres
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space")
-1:12
-(1:12)
-ncol(tdat)
-(1:ncol(tdat))
-?apply
-1:ncol(tdat)
-for (i in 1:ncol(tdat)) {
-paste('region', i)
-}
-for (i in 1:ncol(tdat)) {
-print(paste('region', i))
-}
-regions <- for (i in 1:ncol(tdat)) { print(paste('region', i)) }
-regions <- for (i in 1:ncol(tdat)) { paste('region', i) }
-regions
-for (i in 1:ncol(tdat)) { paste('region', i) }
-regions <- for (i in 1:ncol(tdat)) { paste('region', i) }
-regions
-regions <- list(ncol(tdat))
-?list
-regions <- vector('list', ncol(tdat))
-regions
-?vector
-regions <- vector('character', ncol(tdat))
-regions
-regions[i] <- for (i in 1:ncol(tdat)) { paste('region', i) }
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-regions
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space")
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-21,-0.15))
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.15))
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.16))
-regions <- vector('character', ncol(tdat))
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.165))
-U <- vec.p
-plot(U[,1],U[,2],pch=19,col="darkred",xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-text(U[,1],U[,2],regions,pos=2)
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="Variables in the PC space", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-?var
-dat <- read_delim("~/PycharmProjects/Master/Main/data/genomes/e_coli/FCGRs_for_R", "\t")
-tdat <- t(dat)
-R <- var(tdat)
-#Eigenvalues :
-val.p <- eigen(R)$values
-#Eigenvectors :
-vec.p <- eigen(R)$vectors
-sum(val.p[1:2])/sum(val.p) # % de variance totale expliquée par 2 premiers vecteurs propres
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="PCA of 30 FCGRs (150kb windows)", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-dat <- read_delim("~/PycharmProjects/Master/Main/data/genomes/e_coli/FCGRs_for_R", "\t")
-tdat <- t(dat)
-R <- var(tdat)
-#Eigenvalues :
-val.p <- eigen(R)$values
-#Eigenvectors :
-vec.p <- eigen(R)$vectors
-sum(val.p[1:2])/sum(val.p) # % de variance totale expliquée par 2 premiers vecteurs propres
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="PCA of 30 FCGRs (each from 150kb windows)", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-### PCA analysis of a set of FCGR
-FCGRS <- read_delim("~/PycharmProjects/Master/Main/data/genomes/e_coli/FCGRs_for_R", "\t")
-transposed <- t(FCGRS)
-R <- var(transposed)
-#Eigenvectors :
-U <- eigen(R)$vectors
-# Plot the result
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="PCA of 30 FCGRs (each from 150kb windows)")
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-#b)
-U <- vec.p
-plot(U[,1],U[,2],pch=19,xlab="PC1",ylab="PC2",main="PCA of 30 FCGRs (each from 150kb windows)", xlim=c(-0.21,-0.165))
-regions <- vector('character', ncol(tdat))
-for (i in 1:ncol(tdat)) { regions[i] <- paste('region', i) }
-text(U[,1],U[,2],regions,pos=2)
-U
-args = commandArgs(trailingOnly=TRUE)
-ncol(tdat)
-?png
-FCGRS <- read_delim(args[1], "\t")
-pwd
-??pwd
-?getwd
-getwd()
-FCGRS <- read.table('../files/FCGRs/FCGRs_e_coli', sep = "\t")
-setwd("~/PycharmProjects/Master/Main/scripts")
-FCGRS <- read.table('../files/FCGRs/FCGRs_e_coli', sep = "\t")
-View(FCGRS)
-FCGRS <- read.table('../files/FCGRs/FCGRs_e_coli', sep = "\t", header=TRUE)
